fix: Rename Pill to Chip and Chip to ToggleChip
