+++ src/components/Chip.stories.tsx
@@ -1,6 +1,6 @@
-import { action } from "@storybook/addon-actions";
 import { Meta } from "@storybook/react";
-import { Chip, Css } from "src/index";
+import { Chip } from "src/components/Chip";
+import { Css } from "src/Css";
 
 export default {
   component: Chip,
@@ -9,10 +9,13 @@ export default {
 
 export function DefaultChip() {
   return (
-    <div css={Css.wPx(300).df.fdc.aifs.childGap2.$}>
-      <Chip text="First Last" onClick={action("click")} />
-      <Chip text="Disabled Chip" disabled onClick={action("click")} />
-      <Chip text={"First Last ".repeat(10)} onClick={action("click")} />
+    <div>
+      <div>
+        <Chip text="First Last" />
+      </div>
+      <div css={Css.wPx(300).$}>
+        <Chip text={"First Last ".repeat(10)} />
+      </div>
     </div>
   );
 }
