+++ src/components/Chips.stories.tsx
@@ -1,4 +1,3 @@
-import { action } from "@storybook/addon-actions";
 import { Meta } from "@storybook/react";
 import { Chips } from "src/components/Chips";
 import { Css } from "src/Css";
@@ -11,11 +10,7 @@ export default {
 export function DefaultChips() {
   return (
     <div css={Css.wPx(300).ba.$}>
-      <Chips
-        values={[{ name: "First Last" }, { name: "Second Last" }, { name: "Third Last" }, { name: "Fourth Last" }]}
-        getLabel={(v) => v.name}
-        onRemove={(v) => action(`removed ${v.name}`)()}
-      />
+      <Chips values={["First Last", "Second Last", "Third Last", "Fourth Last"]} />
     </div>
   );
 }
