--- src/components/Pill.tsx
@@ -1,25 +1,0 @@
-import React from "react";
-import { Css, Margin, Only, Xss } from "src/Css";
-import { useTestIds } from "src/utils/useTestIds";
-
-export interface PillProps<X> {
-  text: string;
-  xss?: X;
-}
-
-/** Kinda like a chip, but read-only, so no `onClick` or `hover`. */
-export function Pill<X extends Only<Xss<Margin>, X>>(props: PillProps<X>) {
-  const { text, xss = {} } = props;
-  const tid = useTestIds(props, "pill");
-  return (
-    <span
-      css={{
-        ...Css.dif.aic.br16.sm.pl1.px1.pyPx(2).gray900.bgGray200.$,
-        ...xss,
-      }}
-      {...tid}
-    >
-      {text}
-    </span>
-  );
-}
