+++ src/components/index.ts
@@ -1,12 +1,14 @@
+export * from "src/components/Chip";
+export * from "src/components/Chips";
 export * from "src/components/Table/GridTable";
 export { makeOpenOrCloseCard, makeSpacer, maybeAddCardPadding } from "src/components/Table/nestedCards";
+export * from "src/components/ToggleChip";
+export * from "src/components/ToggleChips";
 export * from "./Alert";
 export { BeamProvider } from "./BeamContext";
 export * from "./Button";
 export * from "./ButtonGroup";
 export * from "./ButtonMenu";
-export * from "./Chip";
-export * from "./Chips";
 export * from "./Copy";
 export * from "./CssReset";
 export * from "./Filters";
@@ -16,8 +18,6 @@ export * from "./Layout";
 export * from "./Modal";
 export * from "./Modal/useModal";
 export { NavLink } from "./NavLink";
-export * from "./Pill";
-export * from "./Pills";
 export { PresentationProvider } from "./PresentationContext";
 export * from "./Stepper";
 export * from "./SuperDrawer";
